Changed token enum to a Token Class
Created a token class. Added a new string parameter "name" to include the name of the Token, as Riley wanted the token text stored in it. (I could be interpreting him wrong, and the parameter could be removed as it seems redundant). 
Created a build method as workaround to the "new" keyword being unable to be used in a nested Class. 
Instead of Token BOOLEAN = new Token(0), it's Token BOOLEAN = build(0, "BOOLEAN");